Move reporting set conversion functions to ProtoConversions
